+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change Log
 
-## 2020-05-10
+## 2020-05-11
 - IR.CreateDevice - added
 - IR.mockAppStart should be called to call IR.EVENT_START and connect all added devices
 - IR.mockResetIr - to reset IR
+- IR.mockAddHost - add hosts that mock devices can connect
 - added jest tests (timers, variables, devices)
 
 ## 2020-01-15